+++ Samples~/Drag And Drop/Scripts/Editor/DragAndDrop.stories.cs
@@ -15,22 +15,22 @@ namespace Unity.AppUI.Sample.Editor
 
         public DragAndDropPage()
         {
-            m_Stories.Add(new StoryBookStory("Main", () =>
-            {
-                var element = new VisualElement();
-                var tree = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(
-                        "Assets/Samples/App UI/1.0.0-pre.3/Drag And Drop/SampleResources/Drag And Drop/DragAndDropUI.uxml");
-                var styleSheet = AssetDatabase.LoadAssetAtPath<ThemeStyleSheet>(
-                    "Assets/Samples/App UI/1.0.0-pre.3/Drag And Drop/SampleResources/Drag And Drop/DragAndDropTheme.tss");
-                tree.CloneTree(element);
-                var root = element.Q<VisualElement>("main-root");
-                var desc = root.Q<Text>("dnd-desc");
-                desc.text += "\nYou can also drag items from others panels of the Editor into the destination list to generate items based on dragged paths.";
-                root.styleSheets.Add(styleSheet);
-                var script = new Unity.AppUI.Samples.DragAndDropSample.DragAndDropSampleScript();
-                script.Start(root);
-                return root;
-            })); 
+            m_Stories.Add(new StoryBookStory("Main", MainStory)); 
+        }
+
+        static VisualElement MainStory()
+        {
+            var element = new VisualElement();
+            var tree = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(AssetDatabase.GUIDToAssetPath("febb74e948b94540999325bf7cf13cd3"));
+            var styleSheet = AssetDatabase.LoadAssetAtPath<ThemeStyleSheet>(AssetDatabase.GUIDToAssetPath("2a9862e95ebae4adaa5eb143ed0b4c98"));
+            tree.CloneTree(element);
+            var root = element.Q<VisualElement>("main-root");
+            var desc = root.Q<Text>("dnd-desc");
+            desc.text += "\nYou can also drag items from others panels of the Editor into the destination list to generate items based on dragged paths.";
+            root.styleSheets.Add(styleSheet);
+            var script = new Unity.AppUI.Samples.DragAndDropSample.DragAndDropSampleScript();
+            script.Start(root);
+            return root;
         }
     }
 }