+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.4] - 2023-11-13
+
+### Added
+
+- Added new *Placement* possibilities in for `AnchorPopup` components. 
+  Now you will be able to anchor a popup inside the anchor element instead of outside of it.
+- Added `zoomChanged` and `scrollOffsetChanged` events to the `Canvas` component.
+- Added `Commanding` documentation page.
+
+### Fixed
+
+- Fixed setting the opacity of `Popover` component at the right frame to avoid to see the content of the popover during the animation.
+
+### Changed
+
+- The `Canvas` component now listens to `PointerDownEvent` in `TrickleDown` phase to avoid to miss the event when the pointer is over a child element.
+
 ## [1.0.0-pre.3] - 2023-11-06
 
 ### Added