+++ Documentation~/whats-new.md
@@ -2,7 +2,7 @@
 uid: whats-new
 ---
 
-# What's New in **1.0.0-pre.3**
+# What's New
 
 This section contains information about new features, improvements, and issues fixed.
 
@@ -10,10 +10,17 @@ For a complete list of changes made, refer to the **Changelog** page.
 
 The main updates in this release include:
 
+### Added
+
+- Added new *Placement* possibilities in for `AnchorPopup` components. 
+  Now you will be able to anchor a popup inside the anchor element instead of outside of it.
+- Added `zoomChanged` and `scrollOffsetChanged` events to the `Canvas` component.
+- Added `Commanding` documentation page.
+
 ### Fixed
 
-- Fixed warning message about Z-axis scale in the `AnchorPopup` component.
+- Fixed setting the opacity of `Popover` component at the right frame to avoid to see the content of the popover during the animation.
 
 ### Changed
 
-- The `Pressable` manipulator now uses the `keepPropagation` property also in its `PointerUpEvent` callback to avoid to propagate the event to the parent element.
+- The `Canvas` component now listens to `PointerDownEvent` in `TrickleDown` phase to avoid to miss the event when the pointer is over a child element.